Reintroduce the adjustment previously done in testcurl.pl so that it copies the CVS checked out curlbuild.h.dist as curlbuild.h for any non-configure target when host system is not running buildconf.bat.
All the curlbuild.h stuff was done taking in consideration that no adjustment
would be needed in non-configure makefiles.

As it is documented, when trying to build on non-configure capable systems or on
systems which for any reason don't run the true configure script, it is required
to have the proper curlbuild.h in place before calling any makefile.

Due to the hardcore memory debugging stuff c-ares enabled debug builds also need
the file in the proper place before attempting to build c-ares.